# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Versatile Timesheet Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects with pagination
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20 }
        - name: status
          in: query
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated list of projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectList"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted

  # ACTIVITY CATEGORIES
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema: { type: boolean }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationList"
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Marked read
...
  /export/employees:
    get:
      operationId: exportEmployees
      tags: [export]
      summary: Export employees as CSV/Excel
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [csv, excel], default: csv }
      responses:
        "200":
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /export/timesheets:
    get:
      operationId: exportTimesheets
      tags: [export]
      summary: Export timesheets as CSV/Excel
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [csv, excel], default: csv }
        - name: dateFrom
          in: query
          schema: { type: string }
        - name: dateTo
          in: query
          schema: { type: string }
      responses:
        "200":
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /export/audit-logs:
    get:
      operationId: exportAuditLogs
      tags: [export]
      summary: Export audit logs as CSV/Excel
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [csv, excel], default: csv }
        - name: dateFrom
          in: query
          schema: { type: string }
        - name: dateTo
          in: query
          schema: { type: string }
      responses:
        "200":
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
        projectCode:
          type: string
        name:
          type: string
        clientId:
          type: string
        clientName:
          type: ["string", "null"]
        status:
          type: string
          enum: [Active, Inactive]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, projectCode, name, clientId, status, createdAt, updatedAt]
    ProjectInput:
      type: object
      properties:
        projectCode:
          type: string
        name:
          type: string
        clientId:
          type: string
        status:
          type: string
          enum: [Active, Inactive]
      required: [projectCode, name, clientId, status]
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        clientId:
          type: string
        status:
          type: string
          enum: [Active, Inactive]
    ProjectList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required: [data, total, page, pageSize]
...
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        relatedId:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, type, title, message, isRead, createdAt]
    NotificationList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total:
          type: integer
        unreadCount:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required: [data, total, unreadCount, page, pageSize]
...